Liten bug
git-svn-id: https://pl3.projectlocker.com/mikaeljorhult/Pizzasugen/svn@46 f4e37ef5-9d55-438e-ab7f-4be2c5091528

+++ trunk/apps/functions.php
@@ -29,7 +29,8 @@ function get_spots($key, $latitude = null, $longitude = null) {
     
     if (count($db_data) > 0) {
       
-      if (strtotime($db_data[0]['time']) < (time() - CACHE_LIFETIME)) {
+      // Only remove the item if lat/lng is provided so that we can do a new search
+      if (strtotime($db_data[0]['time']) < (time() - CACHE_LIFETIME) && is_numeric($latitude) && is_numeric($longitude)) {
           
         $sql = "DELETE FROM `pizzasugen` WHERE `key` = ?";
         $statement = $pdo->prepare($sql);
